+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Version 0.1.8+4(2017/01)
+- Recover rendering performance
+- handle detached state
+
 # Version 0.1.8+4(2016/12)
 - remove polymer tranformer dependency